feat(dev): fast key store for development
Fixes #1614.

We add a command line option to the proxy that makes it use weak
encryption parameters to encrypt files on disk. This allows reasonably
fast encryption even if the proxy is build without code optimization.

The command line option is only available when the proxy is build with
the `unsafe-fast-keystore` feature which is not enabled when building
releases.

Similarly, we also enable the selection of fast encryption for the
remote helper. This is enabled by an environment variable.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>