fix(ui): display branch a commit belongs to
The API now provides the branch a commit belongs to. This partially
reverts #1717.

The issue was not discovered because of disabled type checking due
to #1798. We address this by not using the `Remote` component

The motivation for #1717 was to gracefully handle commits without
branches and not panic. We accomplish this by returning the (possibly
empty) list of all branches and only showing the first branch (if
present) in the UI.

Fixes #1793

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>