fix: remote status indicator vanishing
We fix an issue where the remote status indicator vanished when the user
navigated to the settings page.

The issue was caused by the `remote` store used to hold the status
resetting it’s internal state when the sidebar was re-rendered.

Instead of relying on the remote store with this side-effect we use a
simple writable svelte store.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>