refactor: mutex executor to manage concurrent fetches
We introduce `MutexExecutor` that allows us to abstract over behavior
were we need to abort previous requests issued by the user. We apply
this abstraction in `ui/src/screen/project.ts` where it simplifies the
code significantly.

The abstraction can also be used in other places like
`ui/src/screen/project/source.ts` and the upcoming changes to the
router.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>