refactor: make error code optional
We make the `code` property passed to errors optional. This makes it
easier to add new errors (which happens frequently) because we don’t
need to add it to a global error registry.

Error code’s are beneficial when the cause of the error cannot be
deduced from the message or the stack which rarely is the case, or if we
want to match on a certain type of error.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>