Wait until wallet is connected before fetching orgs
In the tasks that polls for org updates we wait until the wallet has
been connected.  This ensures that we poll as soon as possible after
connecting even if the poll interval is quite large.

To achieve this we introduce the `svelteStore.waitUntil` abstraction.
This abstraction can also be used to implement `waitUnsealed` more
succinctly.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>