refactor: decouple account balance store
We create a separate store for account balances instead of storing that
information in the wallet state.

In addition to the decoupling yielding cleaner code it also fixes some
bugs. Before, we retrieved the balances as part of `loadAccountData`
which delayed updating the connection state. This lead to the UI only
updating some time after we were connected. This also lead to some race
conditions where a disconnect was overriden by a pending
`loadAccountData` call.

In addition to this we also remove `formattedBalance` and inline it.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>