fix: race condition when watching claim updates
We ensure that the `onClaimed` callback is called in the same order as
the events that arrive. Before it was possible for `onClaimed` to be
called for an event Y that occured _after_ the most recent event X
_before_ `onClaimed` was called for `X`. This left the application in a
state where the claim from `X` was considered more recent. This behavior
was triggered by writing tests against a local Ethereum node.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>