docs: describe how to run app against real data
We describe how to run the app from source against real user data. This
should help users that want to build and run the app from source.

We add this to the development doc and the readme to make it more
promentnt. We also move this section to the top of the file to make it
more prominent.

Co-Authored-By: Thomas Scholtes <geigerzaehler@axiom.fm>
Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>