refactor: text input without label
Remove the option to specify a label from TextInput. Instead, the label
is created next to the input. This simplifies the TextInput component.
Since it is not clear how re-usable the label part is (it is only used
in two forms) we also avoid having to adjust the TextInput component
over and over in the future.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>