fix: don’t use stale wallet
Before, some of the functions in `ensRegistrar` would use a style
`Wallet` value because it was set at the module level. Now, we get the
most recent wallet when the functions are called.

Since the functions are inherently stateful, we don’t require functions
to pass down the `environment` parameter and get it from the wallet
instead.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>