dev: use predictable path for test data
For each test we use the same path to store data instead of a new patch
for every run of the same test. This means that we overwrite data from
previous runs. In practice I haven’t found it useful to have data from
previous runs. On the contrary: If there are artifacts from multiple
runs it is difficult to find those of the current run.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>