refactor: move Upstream specific components out of design system
Components that are specific to Upstream or heavily depend on Upstream
code should not be part of the design system, because they're hard to
compose and are unlikely to be reused.

Additionally we consistently use the concept of layouts -- components
that abstract a type of view that is often reused. So far we have a
`<ScreenLayout>` which contains the sidebar and shows a logical view (e.g.
the user profile screen or an org screen) of the application and a
`<ModalLayout>` which shows a full-screen overlay with a modal on top of some
other view.

To make things easier to find and distinguish top-level reusable
components in `ui/App` from views, we add a suffix to components that
are wrapped in either a `<ScreenLayout>` or `<ModalLayout>` with
`Screen` and `Modal` respectively. Components that are related to one or
the other layout now belong in the `ui/App/ScreenLayout` and
`/ui/App/ModalLayout` folders and might later be incorporated directly
into the `<ScreenLayout>` or `<ModalLayout>` components.

Signed-off-by: Rūdolfs Ošiņš <rudolfs@osins.org>