refactor: improve identifier components
Working with identifiers in Upstream @brandonhaslegs and I realised that
we have two distinct use cases:

  - copyable identifiers and
  - identifier links.

Each of them are presented differently to users and have different
behaviour. For this reason we now have `<CopyableIdentifier>` and
`<IdentifierLink>`. We should use `<CopyableIdentifier>` whenever the
entity is not available in the system for direct linking and
`<IdentifierLink>`, when the data is available and we can link to it.

We chose to differentiate the color styling of `<IdentifierLink>`s from
links in the typography system, because screens that have many links to
transaction/commit hashes look too overwhelming in blue.

Signed-off-by: Rūdolfs Ošiņš <rudolfs@osins.org>