fix: force kill test nodes
We forcibly kill the test nodes when they are stopped instead of
shutting them down properly. We have seen the shutdown timing out and
resulting in CI failures when sending `SIGTERM` and we want to avoid
this. We don’t care about state corruption because the node data is not
reused.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>