don’t buffer output for long running subprocesses
We prevent `execa()` from collecting stdout and stderr for processes
that may print a lot of data, in particular the proxy. Otherwise we run
the risk of the Node process to run out of memory.

We also remove the parent process environment from the `env` option.
This is the default behavior for `execa()`.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>