seed: more robust shutdown with task runner
Add `TaskRunner` to provide an extensible platform for managing tasks
and shutdown. The abstraction makes it easy to add additional long
running tasks and ensure proper shutdown.

At the moment this improves the following

* If the tracking task panics `librad` is shut down properly
* The tracking project loop is not just cancelled when a shutdown is
  requested. Instead we finish one iteration.

The logic for `TaskRunner` is very similar to that of `ShutdownRunner`
in the proxy. I opted to not consolidate this for now. We can’t share
code between the two crates yet and there is enough of a difference
between the two implementations.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>