proxy: add option to override seeds
Add a `--seed` option and environment variable that forces the proxy to
use the provided seeds instead of those stored in the configuration.
Using this option will make the seed API read-only and return the seeds
provided on the command line.

We also adjust the environment variable for specifying the default seed.
It now uses the plural `RADICLE_PROXY_DEFAULT_SEEDS` and multiple seeds
can be given by separating them with a comma.

Fixes #2562

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>