tests: work around for failure with unroutable peers
Work around an issue where the protocol shuts down when it tries to
connect to a peer it does not have a route for in the tests. This causes
occasional failures in the tests.

To address the problem, add a default route to the loopback interface
when creating the peer networking stack. This makes all addresses
routable, but the will not arrive unless we have a specific route.

We also refactor the code that setups up the network: We extract
creation of the network stack for a peer into a `create_peer` function.
We also use the `macvlan` virtual interface instead of `veth` so that we
don’t need to deal with two ends.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>