chore: move vagrant cargo target directory
Previously the vagrant target directory was inside the vagrant box which lead
to the file system of the VM to take up a lot of space which can't be
reclaimed by the host. We set the target directory to target/vagrant
so that it resides in the host's file system, yet it doesn't overlap with
the host's target directory.

Signed-off-by: Rūdolfs Ošiņš <rudolfs@osins.org>