run app in dev mode during development
When running the app from the command line during development with `yarn
run electron ./native/index.js` we ensure that the NODE_ENV environment
variable which configures the app is set to "development".

The NODE_ENV variable was properly set when running the app with `yarn
start` but not when running the devnet instantces.

We also enable the `transpileOnly` flag for `ts-node`. This makes the
app start quicker.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>