Add devnet shell command to act as devnet peer
We add a command to the devnet script that allows us to use the radicle
tools in the context of a devnet peer when run as `eval $(yarn run
devnet shell 1)`. We also at an alias for the devnet script to
`package.json` so that we can call it from any directory inside the
repository.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>