proxy: don’t store identity in session
Fixes #2754

The proxy does not store the user identity in it session storage
anymore. There is no need to store the identity because it can always be
retrieved through `librad`.

This requires a change to the `GET /session` endpoint which we use to
detect if we have to create an identity. Before, we used asked the
user to create an identity when no session was stored. The endpoint
retuned a `NoSession` error. Now, we return the `NoSession` error if
there is no key file.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>