migrate keystore and diagnostics endpoins from warp to axum
We integrate `axum` as a replacement for `warp` into `upstream-proxy`
and migrate the `/keystore` and `/diagnostics` endpoints to this
framework.

`axum` has better ergonomics than `warp` and should result in more
readable code and easier to implement handlers.

In addition, we move away from `crate::error::Error` as an umbrella
error to endpoint specific errors.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>