properly open URLs in AppImage
We fix a bug where URLs where not opened properly when running
`radicle-upstream-0.2.13.AppImage URL` while another instance is
running. This bug was a result of the `argv` argument passed to the
`on-second-instance` looking like this:

      argv = [
        '/tmp/.mount_radicl8G9dX0/radicle-upstream',
        '--enable-crashpad',
        '--allow-file-access-from-files',
        '<URL>'
      ]

The extra arguments result in the URL not being in the correct position.
However, `process.argv` for the second instance does not contain the
extra arguments. We use this fact to fix the issue.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>