refactor: localise the daemon code
The radicle-link codebase wants to delete the radicle-daemon code
and continue to work on the linkd/node-lib code as a
replacement. Since radicle-upstream is the main (and only) consumer of
this package, we decided to move the code here for the interim.

The process of the move consisted of:
* Create daemon module under upstream-proxy, copying over all contents
of radicle-link/daemon to here. Fixing up any imports of crate ->
crate::daemon.
* Remove depedency of radicle-daemon in Cargo.toml files
* Update all imports of radicle-daemon -> crate::daemon
* Guard code only used in tests behind #[cfg(test)]
* Remove any unused code that was reported by cargo
* Reword all license headers

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>