chore: remove auth token checks
We want to get ssh-agent authentication to work and getting it play well
with the auth token mechanism is more effort than it's worth. The only
attack vector I see the auth token preventing is an attack from another local
process, which means the local host already might be compromised.

Signed-off-by: Rūdolfs Ošiņš <rudolfs@osins.org>