git fetch refs from all tracked remotes except host peer
We fetch references from Git+HTTP seeds for all peers that we track. We
don’t fetch references belonging to the host peer or track the host
peer. This fixes a bug where the maintainer would show up as a peer that
is being followed in the remote management modal.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>