feat: persist project/seed mapping when tracking projects
This is required for the push/pull integration to work seamlessly
between the `rad` CLI and Upstream.

When we can track a project via `rad clone` or search for it in
Upstream, the seed from which we fetched the project will always be
persisted in the Upstream KV store. This allows us to properly set the
seed URL in the working copy git configuration on subsequent checkouts,
as well as link to the radicle.network web interface from Upstream.

Signed-off-by: Rūdolfs Ošiņš <rudolfs@osins.org>