show domain and port in checkout command
The `rad push` that is part of the command we show to check out a
project now includes the port of the URL, not only the host name.
Seeds with non-standard ports work now.

We also move the extraction logic to the frontend to do as little data
manipulation on the proxy as possible.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>