don’t typecheck on development builds
We don’t run the TypeScript type checker when building with Webpack in
development mode. This makes the builds faster and also allows us to the
app even if some part of it is broken. (This is helpful if the broken
part is unrelated.) Building is also faster.

Without typechecking webpack emits some warnings caused by type
re-exports. We fix these warnings and also enable TypeScript’s
`isolatedModules`  compiler option so that we’re forced to fix these
warnings.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>