Remove standard-version from our release process
Adhering to a formalized commit subject format never caught on and we
also switched to writing human readable release notes.

This removes the standard-version dependency. We now have an extra step during
the release where we write the release notes by hand.

This also removes the requirement to have all commits GPG signed,
because the `rad` CLI tooling (which we use in our day-to-day
development) already uses SSH signatures automatically.

Finally, we change the documentation to relax the requirements on commit
message formatting.

Signed-off-by: Rūdolfs Ošiņš <rudolfs@osins.org>